Internal: enable Next.js ESLint linter
