+++ docs/pages/_document.js
@@ -13,24 +13,6 @@ class GestaltDocument extends Document {
     return (
       <Html lang="en">
         <Head>
-          <script
-            type="text/javascript"
-            src="https://cdn.jsdelivr.net/npm/docsearch.js@2/dist/cdn/docsearch.min.js"
-          />
-          <script async src="https://www.googletagmanager.com/gtag/js?id=UA-12967896-44" />
-          <script
-            // eslint-disable-next-line react/no-danger
-            dangerouslySetInnerHTML={{
-              __html: `
-window.dataLayer = window.dataLayer || [];
-function gtag() {
-  dataLayer.push(arguments);
-}
-gtag('js', new Date());
-gtag('config', 'UA-12967896-44');
-`,
-            }}
-          />
           <link rel="shortcut icon" href="/pinterest_favicon.png" />
         </Head>
         <body>
