+++ docs/pages/whats_new.js
@@ -34,6 +34,7 @@ function Changelog() {
       <Flex alignItems="start" direction="column" gap={4}>
         <Flex gap={4}>
           <Link inline target="blank" href="https://npmjs.org/package/gestalt">
+            {/* eslint-disable-next-line @next/next/no-img-element */}
             <img
               src="https://img.shields.io/npm/v/gestalt.svg?label=gestalt"
               alt="Gestalt NPM package version badge"
@@ -41,6 +42,7 @@ function Changelog() {
           </Link>
 
           <Link inline target="blank" href="https://npmjs.org/package/gestalt-datepicker">
+            {/* eslint-disable-next-line @next/next/no-img-element */}
             <img
               src="https://img.shields.io/npm/v/gestalt-datepicker.svg?label=gestalt-datepicker"
               alt="Gestalt DatePicker NPM package version badge"
