+++ package.json
@@ -19,6 +19,7 @@
     "@babel/preset-flow": "^7.14.5",
     "@babel/preset-react": "^7.13.13",
     "@babel/register": "^7.11.5",
+    "@next/eslint-plugin-next": "^11.1.2",
     "@rollup/plugin-babel": "^5.3.0",
     "@rollup/plugin-commonjs": "^18.0.0",
     "@rollup/plugin-json": "^4.1.0",
